# api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API documentation",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/jobs/{id}": {
      "get": {
        "operationId": "getJobById",
        "tags": [
          "Job"
        ],
        "description": "Get a job by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "queueJobId": {
                      "type": "string",
                      "nullable": true
                    },
                    "htmlContent": {
                      "type": "string",
                      "nullable": true
                    },
                    "url": {
                      "type": "string",
                      "nullable": true
                    },
                    "pdfPath": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PROCESSING",
                        "COMPLETED",
                        "FAILED"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "SYNC",
                        "ASYNC"
                      ]
                    },
                    "finishedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "queueJobId",
                    "htmlContent",
                    "url",
                    "pdfPath",
                    "status",
                    "mode",
                    "finishedAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteJob",
        "tags": [
          "Job"
        ],
        "description": "Delete a job by id",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "queueJobId": {
                      "type": "string",
                      "nullable": true
                    },
                    "htmlContent": {
                      "type": "string",
                      "nullable": true
                    },
                    "url": {
                      "type": "string",
                      "nullable": true
                    },
                    "pdfPath": {
                      "type": "string",
                      "nullable": true
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PROCESSING",
                        "COMPLETED",
                        "FAILED"
                      ]
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "SYNC",
                        "ASYNC"
                      ]
                    },
                    "finishedAt": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "queueJobId",
                    "htmlContent",
                    "url",
                    "pdfPath",
                    "status",
                    "mode",
                    "finishedAt",
                    "createdAt",
                    "updatedAt"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}